Refactored discard deck in to its own file.
