+++ app/src/main/java/com/sq/thed_ck_licker/ecs/components/PlayerComponents.kt
@@ -1,8 +1,6 @@
 package com.sq.thed_ck_licker.ecs.components
 
 import android.util.Log
-import androidx.compose.runtime.MutableIntState
-import androidx.compose.runtime.mutableIntStateOf
 import com.sq.thed_ck_licker.ecs.managers.ComponentManager
 import com.sq.thed_ck_licker.ecs.managers.EntityId
 import com.sq.thed_ck_licker.ecs.managers.hasComponent
@@ -26,25 +24,6 @@ data class DrawDeckComponent @Inject constructor(
 
     fun removeCard(cardId: Int) {
         this.drawCardDeck.remove(cardId)
-    }
-}
-
-
-data class DiscardDeckComponent @Inject constructor(private val discardDeck: MutableList<Int> = mutableListOf<Int>()) {
-    fun getSize(): Int {
-        return this.discardDeck.size
-    }
-
-    fun getDiscardDeck(): MutableList<Int> {
-        return this.discardDeck
-    }
-
-    fun removeCards(cards: List<Int>) {
-        this.discardDeck.removeAll(cards)
-    }
-
-    fun addCard(cardId: Int) {
-        this.discardDeck.add(cardId)
     }
 }
